# apps/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "The public API of the IdeaNest crowdfunding platform, as specified in docs/architecture.md §10.\n\nTwo conventions are worth reading before generating a client against this document. **Money is an object with a string amount** — `{\"amount\": \"599.00\", \"currency\": \"AZN\"}` — and never a JSON number, because a JSON number is an IEEE 754 double and cannot hold what somebody pledged. **Errors are RFC 9457 problem details** carrying a `code` alongside the status; the status says how to behave and the code says what happened, and a client branches on the code.\n\nEndpoints listed under \"Project — public\", discovery, and search answer without an `Authorization` header. Everything else refuses one that is absent.",
    "license" : {
      "name" : "Proprietary"
    },
    "title" : "IdeaNest API",
    "version" : "1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "This deployment"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/v1/projects/{projectId}/fulfilments" : {
      "get" : {
        "operationId" : "fulfilmentOfCampaign",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FulfilmentListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "fulfilment-controller" ]
      }
    },
    "/v1/projects/{projectId}/fulfilments/import" : {
      "post" : {
        "operationId" : "fulfilmentImportTracking",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "text/csv" : {
              "schema" : {
                "type" : "string"
              }
            },
            "text/plain" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FulfilmentImportResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "fulfilment-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FulfilmentImportResponse" : {
        "type" : "object",
        "properties" : {
          "changed" : {
            "type" : "integer",
            "format" : "int32"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RowFailureBody"
            }
          },
          "failed" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rows" : {
            "type" : "integer",
            "format" : "int32"
          },
          "truncated" : {
            "type" : "boolean"
          },
          "unchanged" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "FulfilmentListResponse" : {
        "type" : "object",
        "properties" : {
          "fulfilments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FulfilmentResponse"
            }
          },
          "progress" : {
            "$ref" : "#/components/schemas/FulfilmentProgressResponse"
          }
        }
      },
      "FulfilmentProgressResponse" : {
        "type" : "object",
        "properties" : {
          "backings" : {
            "type" : "integer",
            "format" : "int64"
          },
          "delivered" : {
            "type" : "integer",
            "format" : "int64"
          },
          "preparing" : {
            "type" : "integer",
            "format" : "int64"
          },
          "returned" : {
            "type" : "integer",
            "format" : "int64"
          },
          "shipped" : {
            "type" : "integer",
            "format" : "int64"
          },
          "untouched" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "FulfilmentResponse" : {
        "type" : "object",
        "properties" : {
          "carrier" : {
            "type" : "string"
          },
          "deliveredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "pledgeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "shippedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PREPARING", "SHIPPED", "DELIVERED", "RETURNED" ]
          },
          "trackingNumber" : {
            "type" : "string"
          },
          "trackingUrl" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "RowFailureBody" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "line" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "pledgeId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "A fifteen-minute access token from POST /v1/auth/login or /v1/auth/refresh. The refresh token is not sent this way: it lives in an httpOnly cookie on web and in secure storage on mobile.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}